+++ src/webviews/components/Icons/GitPullRequestIcon.tsx
@@ -1,21 +1,33 @@
+import {
+  GIT_PULL_REQUEST_CREATE_ICON_PATH,
+  GIT_PULL_REQUEST_ICON_PATH,
+} from "./gitPullRequestIconPath"
+
 type GitPullRequestIconProps = {
   size?: number
   style?: React.CSSProperties
+  className?: string
+  variant?: "create" | "view"
 }
 
 export function GitPullRequestIcon(props: GitPullRequestIconProps) {
-  const { size = 14, style } = props
+  const { size = 14, style, className, variant = "create" } = props
+  const path = variant === "view" ? GIT_PULL_REQUEST_ICON_PATH : GIT_PULL_REQUEST_CREATE_ICON_PATH
 
   return (
     <svg
+      className={className}
       width={size}
       height={size}
+      style={style}
       viewBox="0 0 16 16"
       fill="currentColor"
-      style={style}
+      role="img"
+      focusable="false"
       aria-hidden="true"
+      xmlns="http://www.w3.org/2000/svg"
     >
-      <path d="M7.177 2.672A2.25 2.25 0 0 1 9.75 4.5v.75a.75.75 0 0 1-1.5 0v-.75a.75.75 0 0 0-.75-.75h-2.5a.75.75 0 0 0-.75.75v.75a.75.75 0 0 1-1.5 0v-.75A2.25 2.25 0 0 1 6.177 2.672h1ZM4.5 7.5a.75.75 0 0 1 .75-.75h5.5a.75.75 0 0 1 0 1.5h-5.5a.75.75 0 0 1-.75-.75ZM4 10.25a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm7.5 0a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5ZM8 4.25a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5Z" />
+      <path fillRule="evenodd" clipRule="evenodd" d={path} />
     </svg>
   )
 }