+++ src/webviews/entries/issue.tsx
@@ -1,0 +1,4 @@
+import { mountWebview } from "../bootstrap"
+import IssueWebview from "../views/IssueWebview/IssueWebview"
+
+mountWebview(<IssueWebview />)