+++ src/webviews/entries/settings.tsx
@@ -1,0 +1,4 @@
+import { mountWebview } from "../bootstrap"
+import { SettingsWebview } from "../views/SettingsWebview/SettingsWebview"
+
+mountWebview(<SettingsWebview />)