+++ src/webviews/entries/startWork.tsx
@@ -1,0 +1,4 @@
+import { mountWebview } from "../bootstrap"
+import { StartWorkWebview } from "../views/StartWorkWebview/StartWorkWebview"
+
+mountWebview(<StartWorkWebview />)